+++ tests/test_plugins/test_adjoint.py
@@ -51,7 +51,6 @@ TMP_PATH = None
 FWD_SIM_DATA_FILE = "adjoint_grad_data_fwd.hdf5"
 SIM_VJP_FILE = "adjoint_sim_vjp_file.hdf5"
 RUN_FILE = "simulation.hdf5"
-NUM_PROC_PARALLEL = 2
 
 EPS = 2.0
 SIZE = (1.0, 2.0, 3.0)
@@ -115,7 +114,6 @@ def run_emulated_bwd(
     folder_name: str,
     callback_url: str,
     verbose: bool,
-    num_proc: int = NUM_PROC_PARALLEL,
 ) -> JaxSimulation:
     """Runs adjoint simulation on our servers, grabs the gradient data from fwd for processing."""
 
@@ -135,9 +133,7 @@ def run_emulated_bwd(
     grad_data_adj = jax_sim_data_adj.grad_data_symmetry
 
     # get gradient and insert into the resulting simulation structure medium
-    sim_vjp = jax_sim_data_adj.simulation.store_vjp(
-        grad_data_fwd, grad_data_adj, grad_eps_data_fwd, num_proc=num_proc
-    )
+    sim_vjp = jax_sim_data_adj.simulation.store_vjp(grad_data_fwd, grad_data_adj, grad_eps_data_fwd)
 
     # write VJP sim to and from file to emulate webapi download and loading
     sim_vjp.to_file(str(TMP_PATH / SIM_VJP_FILE))
@@ -198,7 +194,6 @@ def run_async_emulated_bwd(
             folder_name=folder_name,
             callback_url=callback_url,
             verbose=verbose,
-            num_proc=NUM_PROC_PARALLEL,
         )
         sim_vjps_orig.append(sim_vjp)